feat: migrate test_api file
