+++ lms/djangoapps/discussion/rest_api/tests/utils.py
@@ -15,6 +15,7 @@ from PIL import Image
 from pytz import UTC
 
 from lms.djangoapps.discussion.django_comment_client.tests.mixins import MockForumApiMixin
+from openedx.core.djangoapps.django_comment_common.comment_client.utils import CommentClientRequestError
 from openedx.core.djangoapps.profile_images.images import create_profile_images
 from openedx.core.djangoapps.profile_images.tests.helpers import make_image_file
 from openedx.core.djangoapps.user_api.accounts.image_helpers import get_profile_image_names, set_has_profile_image
@@ -625,7 +626,10 @@ class ForumMockUtilsMixin(MockForumApiMixin):
         self.set_mock_side_effect('update_thread', make_thread_callback(thread_data))
 
     def register_get_thread_error_response(self, thread_id, status_code):
-        self.set_mock_return_value('get_thread', Exception(f"Error {status_code}"))
+        self.set_mock_side_effect(
+            'get_thread',
+            CommentClientRequestError(f"Thread does not exist with Id: {thread_id}")
+        )
 
     def register_get_thread_response(self, thread):
         self.set_mock_return_value('get_thread', thread)
@@ -653,7 +657,10 @@ class ForumMockUtilsMixin(MockForumApiMixin):
         )
 
     def register_get_comment_error_response(self, comment_id, status_code):
-        self.set_mock_return_value('get_parent_comment', Exception(f"Error {status_code}"))
+        self.set_mock_side_effect(
+            'get_parent_comment',
+            CommentClientRequestError(f"Comment does not exist with Id: {comment_id}")
+        )
 
     def register_get_comment_response(self, response_overrides):
         comment = make_minimal_cs_comment(response_overrides)
